requirements.txt
fastapi
uvicorn[standard]
python-dotenv
websockets
mcp[cli]
aws_sdk_bedrock_runtime
smithy-aws-core
psycopg2-binary
squareup